recycles old HL_History / works as expected
I changed the score so that it would say how much the user had got right and how many they had got wrong. (instead of saying won or lost) and also changed the quiz scores so that instead of saying how many guesses the user used it would show the score.