Fixed code now works as expected :)
added level choice and rounds to questions ( equations ). Now this code produces a quiz with questions that can play with any level (easy, medium, hard). continuous mode is available  if user press <enter>. Emergency Exit  also works in this code. After the amount of rounds (questions) asked is played then a temporary score total is shown at the end of the game. The code is looped so that the user can play again after every quiz.